Accept both bytes and bytearray for writeCODE
Make writeCODE work whether data is a bytes, str, or bytearray object. This
means we have to change any loops through data, because elements of bytes are
bytes and elements of bytearray are ints. For the write, simply pass the whole
thing to pyserial.write instead of looping through it. For verifying, check the
whole data instead of looping through it. This breaks the functionality that
the verify function tells the exact offset of the data error. The retry code is
also gone, but this already didn't really retry anything.

The motivation for this change was that CCHEXFile._loadbin returns a str, and
CCHEXFile._loadhex returns a bytearray. This change makes cc_write_flash.py
work with .bin files.